Enables CPUAPP variant and configures peripherals
Adds a cpuapp variant for the nrf54l15 SoC.

This commit configures the device tree and Kconfig files
to enable the PWM, GPIO, and ADC peripherals. It also sets
up memory partitions for mcuboot.

The UART30 and SPI00 configurations are removed.
The pwmleds configuration is replaced with a pwm_beep configuration.